refactor(mobile): tratamento de erros e reutilização do fetch de matérias e aulas
